% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_centroid.R
\name{estimate_centroid}
\alias{estimate_centroid}
\title{Estimate Phenological Centroid (DOBY)}
\usage{
estimate_centroid(model, newdata, n_sims = 1000, sim_mat = NULL)
}
\arguments{
\item{model}{A fitted mgcv GAM/BAM object}

\item{newdata}{A data.frame with prediction grid (365 days)}

\item{n_sims}{Integer, number of posterior draws}

\item{sim_mat}{Optional pre-computed simulation matrix}
}
\description{
Estimate Phenological Centroid (DOBY)
}
