% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_duration.R
\name{estimate_duration}
\alias{estimate_duration}
\title{Estimate Phenological Active Duration}
\usage{
estimate_duration(
  model,
  newdata,
  n_sims = 1000,
  sim_mat = NULL,
  threshold_fraction = 0.05
)
}
\arguments{
\item{model}{A fitted mgcv GAM/BAM object}

\item{newdata}{A data.frame with prediction grid (365 days)}

\item{n_sims}{Integer, number of posterior draws}

\item{sim_mat}{Optional pre-computed simulation matrix}

\item{threshold_fraction}{The density threshold as a fraction of peak maximum (default 0.05)}
}
\description{
Estimate Phenological Active Duration
}
