% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metric_gradient.R
\name{plot_metric_gradient}
\alias{plot_metric_gradient}
\title{Plot Metric Gradient}
\usage{
plot_metric_gradient(
  model,
  dataset_for_quantiles,
  metric = "centroid",
  x_axis = "latitude",
  temperature_col = "TMAM",
  hold_constant = "mean",
  n_sims = 1000,
  add = FALSE,
  color = "#4575b4"
)
}
\description{
Plot Metric Gradient
}
