% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phenology_curve.R
\name{plot_phenology_curve}
\alias{plot_phenology_curve}
\title{Plot Phenology Curve}
\usage{
plot_phenology_curve(
  model,
  dataset_for_quantiles,
  latitude = "mean",
  temperature = "mean",
  n_sims = 1000,
  log = TRUE,
  add = FALSE,
  color = "#d73027"
)
}
\description{
Plot Phenology Curve
}
